[추가] 테스트코드 파일을 분리하여 정돈합니다.
