+++ KaleubSDK/API/Room/List/RoomListRequest.swift
@@ -23,7 +23,7 @@ struct RoomListRequest: Request, Respondable {
         var urlRequest = URLRequest(url: url)
         urlRequest.httpMethod = method.stringValue
         // NOTE: - UserToken을 넣어야 합니다.
-        urlRequest.addValue(userToken, forHTTPHeaderField: "Authorization")
+        urlRequest.addValue("Bearer \(userToken)", forHTTPHeaderField: "Authorization")
         return urlRequest
     }
 }
