+++ KaleubSDKTests/APIClientTests/APIClientTests.Room.swift
@@ -1,0 +1,34 @@
+//
+//  APIClientTests.Room.swift
+//  KaleubSDKTests
+//
+//  Created by Hamlit Jason on 2022/06/28.
+//
+
+import XCTest
+@testable import KaleubSDK
+
+extension APIClientTests {
+    func test_createRoom() {
+        let request = RoomListRequest()
+        
+        let expectation = XCTestExpectation()
+        expectation.expectedFulfillmentCount = 1
+        
+        apiClient?.createRoom(title: "첫번쨰방", password: "123456") { result in
+            switch result {
+            case .success:
+                XCTAssert(true)
+            case .failure(let error):
+                if let error = error as? APIClient.APIError {
+                    XCTFail(error.localizedDescription)
+                } else {
+                    XCTFail(error.localizedDescription)
+                }
+            }
+        }
+        
+        expectation.fulfill()
+    }
+
+}
